tour culling round 2

+++ svinbin/reports/serializers.py
@@ -53,6 +53,6 @@ class StartDateEndDateSerializer(serializers.Serializer):
     
 class TotalWeightsSerializer(serializers.Serializer):
     total_quantity = serializers.IntegerField()
-    total_avg = serializers.FloatField(decimal_places=2)
-    total_total_weight = serializers.IntegerField(decimal_places=2)
-    total_avg_age = serializers.IntegerField(decimal_places=2)
+    total_avg = serializers.DecimalField(max_digits=10, decimal_places=2)
+    total_total_weight = serializers.DecimalField(max_digits=10, decimal_places=2)
+    total_avg_age = serializers.DecimalField(max_digits=10, decimal_places=2)
